Author: Bleep <tomh@inxpress.net> Log message: Merge in changes from beta
git-svn-id: https://undernet-ircu.svn.sourceforge.net/svnroot/undernet-ircu/ircu2/trunk@37 c9e4aea6-c8fd-4c43-8297-357d70d61c8c

+++ ircd/m_kill.c
@@ -441,7 +441,7 @@ int m_kill(struct Client *cptr, struct Client *sptr, int parc, char *parv[])
     if (EmptyString(path))
       return need_more_params(sptr, parv[0], "KILL");
 
-    if (strlen(path) > (size_t)TOPICLEN)
+    if (strlen(path) > TOPICLEN)
       path[TOPICLEN] = '\0';
   }
